feat(shortener): add support for memphis to send analytics data
This commit adds support for memphis to send analytics data. It includes
changes to the .example.env file to add MEMPHIS_ENABLE, MEMPHIS_HOST,
MEMPHIS_USERNAME, and MEMPHIS_PASSWORD environment variables. It also
includes changes to the config factory to add the memphis.enable
property to the MemphisConfig interface. The memphis module is added to
the analytics module, and the analytics producer is moved to the
analytics module. The shortener controller is updated to import the
analytics module. Finally, the memphis service mock is added to the
memphis module.